requirements.txt
Pillow
psycopg2-binary
python-decouple
python-slugify
wheel
djangorestframework
django-filter
drf-yasg
pywhatkit